pyproject.toml
[project]
name = "agh-ai-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pre-commit>=4.2.0",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "xenon>=0.9.3",
]
